fixed issue xhr request not reporting errors
- removed module from modules bucket in the compilation step for
modules that are the primary target. You can have multiple module
definition in a single file and only one is the one that corresponds to
the module id loaded. that module is removed from the generic buckets
in the compilation step.
- also adjusted unit tests to follow API changes in bit loader where
passing in an array of modules to be loaded now returns an array of
modules.